+++ bebop-lang/bench/vs_rust/construct_parity.sh
@@ -98,6 +98,7 @@ for f in "${DIR%/}/neg"/*.bp; do
     c29_emptybody) EXPECT=COMPILEFAIL:97;;
     c37_arenafull) EXPECT=RUNFAIL:80;;
     c38_frameheap) EXPECT=RUNFAIL:81;;
+    c39_fnmatch) EXPECT=COMPILEFAIL:99;;
     *) EXPECT="";;
   esac
   out="${BEBOP_TMP:-/tmp/opencode}/${b}_test.bin"